Add new document:upsert API action (#1898)
# Description

Fix #1895 

Add a new `document:upsert` API action, able to either apply partial changes to a document, or to create a new document on the fly.

# Query syntax

```js
{
  "index": "<index>",
  "collection": "<collection>",
  "controller": "document",
  "action": "update",
  "_id": "<documentId>",
  "body": {
    "changes": {
      // document partial changes
    },
    "default": {
      // optional: document fields to add to the changes if the document
      // is created
    }
  }
}
```

The body part of the query contains 2 objects:
* `changes` (required): the partial changes to apply to existing documents. 
* `default` (optional): optional fields to add to the new document if it doesn't already exist

So, this route does this:
* if the document already exists, this API action behaves exactly as `document:update`: the content of `body.changes` is applied to the existing document (and the content of `body.default` is ignored)
* if the document doesn't exist, a new one is created, containing both the content of `body.default` and `body.changes`, IN THAT ORDER (meaning that if a field exists both in changes and in default, the one in `changes` supersedes the default one)

# API response

This route takes the same options as `document:update`: `source` (boolean), `retryOnConflict` (number), `refresh` (string)

But it behaves a bit differently with the `source` option: by default (`source = false`), it doesn't return any `_source` property in the response.

Also, compared to `document:update`, a new property is available in the response: `_created` (boolean).
If `true`, the document didn't exist and was created. It is set to `false` otherwise.

# Metadata

The content of `_kuzzle_info` varies depending on whether the document was created or updated, matching what is currently found in documents if `document:create` or `document:update` was called, respectively.

# Realtime notifications

Realtime notifications triggered match the status of the document: if it was created, a "new document" notification is fired, otherwise a "document updated" is sent to subscribers.