Make kuzzle instance a singleton (#1911)
# Description


Kuzzle instances are now singletons. And they are put in the global space.
This PR removes usages of `this.kuzzle` from everywhere, removing a lot of clutter.

# Why?

I need to expose the mutex class to plugins and applications. To do this properly, I need to convert the mutex object to TS.
But... the mutex object needs the Kuzzle object. And we don't want plugins/apps user to provide one, obviously.

So, I had 2 possibilities: write yet another TS wrapper that will inject the kuzzle object to the mutex object. Or get rid of that kuzzle object once and for all, simplifying the code.

Since having `this.kuzzle` everywhere was a pain point, I decided for the latter solution (mutex for plugins/apps will come in another PR)

# Other changes

Fix a bug in the graceful shutdown, where the number of remaining requests was never refreshed, thus preventing Kuzzle to shut down properly if there were pending requests.