Hotfix: Restore API keys when the cache is wiped clean (#1941)
# Description

This PR solves 2 problems.

First, API keys are stored in Elasticsearch, and synchronized in Redis. They are supposed to be loaded into Redis during the bootstrap of the internal indexes, which are run only once during the lifetime of Kuzzle: when it's started for the first time. 
If the Redis cache is wiped, or when Kuzzle is started for the first time with an ES storage restored from a snapshot, Kuzzle skips the internal index bootstrap, and API keys aren't synchronized in Redis, making them unusable.

Second, loading API keys in the internal index bootstrapper doesn't work. This wasn't visible because the bootstrap only occured when Kuzzle is started for the first time, on a blank ES storage. But the API key loader uses the Token Repository, which isn't initialized until after the internal indexes are bootstrapped.

# Solution

The API keys synchronization is now the responsibility of the Token Repository, and a different bootstrap key is used. Contrary to the one used by the internal index bootstrap, this bootstrap key is stored in Redis.
This way, if the cache is wiped, or if Kuzzle is started from a restored ES storage, the Token Repository correctly restores the API keys in the cache, and this is done only once, whatever the number of Kuzzle nodes in a cluster environment.

# Why is this a hotfix?

Because one of our clients is currently experiencing this issue, and it's a blocking one.