Upgrade script: URL validation didn't accept non-FQDN addresses (#1983)
## What does this PR do ?

Fixes #1978 
Fixes #1979

When configuring Elasticsearch connections for the data migration, URLs that aren't fully qualified aren't accepted.

On the other hand, URLs without a protocol or port are accepted by the script, but not by the ES client.

This PR fixes this by harmonizing URL validation with what is expected by the ES client. Now, URLs entered by users must have a protocol set (http or https), a port (since ES isn't exposed through the port 80), and URLs that aren't fully-qualified are accepted (e.g. `http://localhost:9200`)

Also, the error message has been updated to act as an helper, providing an example of a valid URL.

### How should this be manually tested?

Run the upgrade script. When asked about a source Elasticsearch URL, try the following:

| URL | Result (before) | Result (after) |
| --- | --- | --- |
| `localhost:9200` | refused | refused |
| `http://localhost:9200` | refused | accepted |
| `10.0.0.1:9200` | accepted by the script, refused by the ES client | refused |
| `10.0.0.1` | accepted by the script, refused by the ES client | refused |
| `http://10.0.0.1` | accepted by the script, cannot connect to ES | refused |
| `http://10.0.0.1:9200` | accepted | accepted |

### Other changes

The `realtime` configuration entry was marked as deprecated, and so to be removed, by the upgrade script.
Though it's true that this option is deprecated (it's used to be able to use the PCRE engine with Koncorde instead of RE2, which is NOT RECOMMENDED for security reasons), it's still accepted by Kuzzle v2, and thus can be used.

This entry has been removed from the `deprecated` list, since this makes the upgrade script complain with our own default configuration file, which sounds like a bug to me.

### Boyscout

Converted a few scripts to our latest coding standards.