Allows dynamic pipe registration (#1996)
## What does this PR do ?

Allows to dynamically register pipes.

When the application is already started, `app.pipe.register` can be used with the `dynamic: true` option to dynamically register a pipe.  
The method will return a `pipeId` which can be then used later to unregister the pipe with `app.pipe.unregister`.

Defining dynamic pipes at return need to maintain a state of the registered pipes so new nodes joining the cluster can also register the same pipes.  
This dynamic state management is not handled by Kuzzle so developers should implement their own state management routines.

This feature is meant to be left undocumented since it's quite and advanced usage, and we didn't have other use cases than our Workflows plugin.

### How should this be manually tested?

  - Step 1 : Run functional tests

### Other

 - Expose a usable version of Koncorde