feat: allow overriding configuration properties with JSON strings (#2012)
Properties of the kuzzlerc configuration file can be overridden using environment variables. This is a native property of the rc module that we use to load configuration values.

Problem is: rc does only handle scalar values. Meaning that users cannot update arrays using environment variables.

This PR proposes to solve this issue by allowing users to pass JSON strings in their environment variable: with this PR, Kuzzle will parse those strings as a post-processing to the rc configuration loading.

This is done by passing the *json: prefix to a JSON string (see the updated documentation for examples).