Cluster: die if unable to refresh the ID card in time (#2051)
# Description

Each cluster node maintains its own ID card in redis, allowing other nodes to discover the node, and connect to it.

ID cards are temporary redis key, meaning that they disappear after a time. It's intended: if a node crashes, its ID card disappears and future nodes won't try to connect to it.

But if a node is too slow to refresh its own ID card, then the key disappears: future nodes don't know that this node exists, they never form a cluster, and this makes the cluster inconsistent.

With this PR, each time a node refreshes its ID card, it checks that the refresh was successful: Redis returns 1 if the key is refreshed, and 0 if the key doesn't exist. In the latter case, the node now detects it, and shuts itself down.