Access Logs: request responses may contain unserializable objects
# Description

Since Kuzzle 2.12.0, access logs are isolated in a worker thread, to free resources for Kuzzle's main thread.
This gives a huge performances boost to Kuzzle.

Now, this requires the KuzzleRequest object to be serialized and passed over to the worker thread, so that it can, in turn, compute access logs information and send them to the winston transport(s).

Problem is: not everything can be serialized that way. Trying to pass unserializable content to a worker thread results in a `DataCloneError` from Node.js, which makes Kuzzle crash.

In that sense, serializing a KuzzleRequest is generally safe, apart from the RequestResponse sub-object, which contains whatever plugin developers set to it.
Kuzzle makes sure that request responses are stringifiable, but... objects that can be stringified might not be serializable.

This PR solves that issue by removing the request response from the object passed to the access logger thread. It is not needed anyway, apart from the content size that needs to be logged: so the main thread now takes care of computing it, and passing the result to the thread.

# How to reproduce

Write a custom plugin, extending the API. 
In the plugin's API response, return an object containing a Javascript function (functions are ignored during stringification, but trigger a DataCloneError when attempting to serialize them).

Plugin Example:

```js
class Foo {
  constructor() {
  }

  async init (config, context) {
    this.config = config;
    this.context = context;
    this.sdk = context.accessors.sdk;

    this.api = {
      foo: {
        actions: {
          bar: {
            handler: rq => this.ohnoes(rq),
          }
        },
      },
    };
  }

  async ohnoes (rq) {
    return { 
      foo: 'bar', 
      bar: function () { console.log('ohnoes'); } 
    };
  }
}
```

Load the plugin, call its API route, and watch Kuzzle die.