Faster collection truncating with deleteByQuery (#2107)
When truncating a collection, Kuzzle delete and then re-create the underlying Elasticsearch indice.   
However indice creation is not parallelizable inside an ES cluster so if we truncate many collections it will be quite slow when the deleteByQuery action is completely parallel. 

This PR change `collection:truncate` to use `deleteByQuery` which is almost always faster.