Increase ID card expiration delay (#2183)
## What does this PR do ?

Increase the ID Card expiration delay. Now the refresh delay is multiplied by 4 so by default it's 8 seconds TTL.

I noticed that Node.js takes a lot of time to start the worker and because of that sometime when the worker start and try to refresh the ID card it's already expired.