Release 2.14.16: Store ID Card's keys in a Redis set to speed up handshake (#2190)
Each node ID Card is stored in Redis so when a new node starts, it can retrieve other nodes ID Cards to connect to them. The SCAN command was used to retrieve keys corresponding to a node ID Card.

When Redis database is growing, the time needed is also growing so the handshake procedure can end up with a timeout and preventing new nodes to join the cluster.

The cluster now use an Redis Set to store the list of ID Card's keys, when a new node starts, it will store it's own key inside the set, then get it to retrieve the name of each ID Card's keys to finally retrieve the ID Cards themselves.
If some ID Cards can not be retrieved, it means they expired so the new node will clean their keys from the set.