Remove unicode property names from KuzzleRequest (#2216)
In the KuzzleRequest object, most of the property were using a special Unicode character to prevents user from modifying them (private attribute emulation).

This was preventing Koncorde to match content of the KuzzleRequest object directly.

Those properties now have ASCII names. If an user want to edit one of them he can but Kuzzle may not works after that.