Release 2.16.3: Allows to match KuzzleRequest with Koncorde (#2231)
## What does this PR do ?

Revert https://github.com/kuzzleio/kuzzle/pull/2216

Having getter setter + invisible unicode properties allows to have a safe Kuzzle object. Despite the fact that the developers are responsible for not messing with the `KuzzleRequest` object, the fact that it was not editable allowed more stability in Kuzzle applications.

The whole point is was to be able to match KuzzleRequest object in Koncorde filters. There is only 2 solutions:
 - rename unicode properties so we can match them (e.g. `{ equals:  { "input.args.collection": "devices" } }`)
 - modify Koncorde to handle getters

The first solution leads to stability issues for Kuzzle.
The second solution results in a massive performance drop for Koncorde.

This PR adds a `request.pojo()` method that returns a POJO that can be matched with Koncorde, without unicode properties