Allows to register realtime subscriptions from the backend (#2440)
Realtime subscription can be added on a connection directly from the backend. This allows to restrict the usage of the realtime:subscribe and to add realtime subscription from a custom API action.

The Backend.subscription.add method takes the same arguments as the realtime:subscribe API action but allows to register the subscription on any connection.